requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
aiohappyeyeballs==2.3.5
    # via aiohttp
aiohttp==3.10.2
    # via
    #   langchain
    #   voyageai
aiolimiter==1.1.0
    # via voyageai
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anthropic==0.32.0
    # via -r requirements.in
anyio==4.4.0
    # via
    #   anthropic
    #   httpx
    #   openai
attrs==24.2.0
    # via aiohttp
boto3==1.34.157
    # via cohere
botocore==1.34.157
    # via
    #   boto3
    #   s3transfer
certifi==2024.7.4
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
cohere==5.6.2
    # via -r requirements.in
diskcache==5.6.3
    # via -r requirements.in
distro==1.9.0
    # via
    #   anthropic
    #   openai
faiss-cpu==1.8.0.post1
    # via -r requirements.in
fastavro==1.9.5
    # via cohere
filelock==3.15.4
    # via huggingface-hub
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
fsspec==2024.6.1
    # via huggingface-hub
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via
    #   anthropic
    #   cohere
    #   openai
httpx-sse==0.4.0
    # via cohere
huggingface-hub==0.24.5
    # via tokenizers
idna==3.7
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jiter==0.5.0
    # via
    #   anthropic
    #   openai
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==0.2.13
    # via -r requirements.in
langchain-core==0.2.30
    # via
    #   langchain
    #   langchain-text-splitters
langchain-text-splitters==0.2.2
    # via langchain
langsmith==0.1.99
    # via
    #   langchain
    #   langchain-core
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
mypy==1.11.1
    # via -r requirements.in
mypy-extensions==1.0.0
    # via mypy
numpy==1.26.4
    # via
    #   faiss-cpu
    #   langchain
    #   pandas
    #   pandas-stubs
    #   voyageai
openai==1.40.1
    # via -r requirements.in
orjson==3.10.7
    # via langsmith
packaging==24.1
    # via
    #   faiss-cpu
    #   huggingface-hub
    #   langchain-core
pandas==2.2.2
    # via -r requirements.in
pandas-stubs==2.2.2.240807
    # via -r requirements.in
parameterized==0.9.0
    # via cohere
pydantic==2.8.2
    # via
    #   -r requirements.in
    #   anthropic
    #   cohere
    #   langchain
    #   langchain-core
    #   langsmith
    #   openai
pydantic-core==2.20.1
    # via pydantic
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   pandas
pytz==2024.1
    # via pandas
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   langchain
    #   langchain-core
regex==2024.7.24
    # via tiktoken
requests==2.32.3
    # via
    #   cohere
    #   huggingface-hub
    #   langchain
    #   langsmith
    #   tiktoken
    #   voyageai
ruff==0.5.7
    # via -r requirements.in
s3transfer==0.10.2
    # via boto3
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anthropic
    #   anyio
    #   httpx
    #   openai
sqlalchemy==2.0.32
    # via langchain
tenacity==8.5.0
    # via
    #   langchain
    #   langchain-core
    #   voyageai
tiktoken==0.7.0
    # via -r requirements.in
tokenizers==0.20.0
    # via
    #   anthropic
    #   cohere
tqdm==4.66.5
    # via
    #   huggingface-hub
    #   openai
types-pytz==2024.1.0.20240417
    # via pandas-stubs
types-requests==2.32.0.20240712
    # via cohere
types-setuptools==71.1.0.20240806
    # via -r requirements.in
types-tqdm==4.66.0.20240417
    # via -r requirements.in
typing-extensions==4.12.2
    # via
    #   anthropic
    #   cohere
    #   huggingface-hub
    #   langchain-core
    #   mypy
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
tzdata==2024.1
    # via pandas
unidecode==1.3.8
    # via -r requirements.in
urllib3==2.2.2
    # via
    #   botocore
    #   requests
    #   types-requests
voyageai==0.2.3
    # via -r requirements.in
yarl==1.9.4
    # via aiohttp
